Move FixedPointSolver into a separate module, add documentation
