+++ ACKNOWLEDGEMENTS
@@ -103,10 +103,12 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 The FixedPointSolver class in tlm_adjoint/equations.py derives tangent-linear
 and adjoint information using the approach described in [1, 2].
 
-[1]  J. C. Gilbert, "Automatic differentiation and iterative processes",
-     Optimization Methods and Software, 1(1), pp. 13--21, 1992
-[2]  B. Christianson, "Reverse accumulation and attractive fixed points",
-     Optimization Methods and Software, 3(4), pp. 311--326, 1994
+[1] Jean Charles Gilbert, 'Automatic differentiation and iterative processes',
+    Optimization Methods and Software, 1(1), pp. 13--21, 1992,
+    doi: 10.1080/10556789208805503
+[2] Bruce Christianson, 'Reverse accumulation and attractive fixed points',
+    Optimization Methods and Software, 3(4), pp. 311--326, 1994,
+    doi: 10.1080/10556789408805572
 
 ===============================================================================
